YES! This fixes #8863, which has been on my plate forever. Woot!
As it turns out, using a private API here worked in 10.4, but in 10.5, it set the error condition on the apple event.
In the future, a rewriting of the applescript stuff to take advantage of the new 10.5 features is definitely in order.

--HG--
extra : convert_revision : svn%3A32148f9c-d742-0410-bc91-dd12f8e4a632/trunk%4025251