Provide the Twitter service icon as a temporary serverside icon. When an {{{AIListContact}}} doesn't have an icon (yet), we need to provide something to the WKMV so it has an image to insert (and later update). Otherwise we will never display the new icon in the chat.
Twitter contacts are __guaranteed__ to have a serverside icon, so I don't see a problem in doing this for Twitter contacts. As soon as the actual one is updated, we'll replace this icon with it. Other services don't meet the same criteria, so implementing this as a global change would cause lots of service icons in chats.

--HG--
extra : convert_revision : svn%3A32148f9c-d742-0410-bc91-dd12f8e4a632/trunk%4026057