When logging XML, don't eliminate significant spaces. Instead, replace them with an en-space, U+2002.
When a space is in-between two elements, i.e. {{{<a>…</a> <a>…</a>}}}, it would be compressed to blank by the logger, and subsequently when read in. Eliminate this problem by replacing single-space text additions with the en-space.

--HG--
extra : convert_revision : svn%3A32148f9c-d742-0410-bc91-dd12f8e4a632/trunk%4026387