Store flags and aliases for chats in the appropriate chats. Creates a new AIListContact cell for displaying in group chats, which prefixes the user with the character representing their status. As of yet, they aren't sorted by this order.
Adds a new method to the contact controller to set the UID for a contact. Since it stores contacts based on their internal ID, which is a product of the UID, merely renaming on the AIListContact level wasn't enough; we also had to change the contact controller's value.

--HG--
extra : convert_revision : svn%3A32148f9c-d742-0410-bc91-dd12f8e4a632/trunk%4026607