Always try and reposition the contact list if we think it should be. Fixes #12716.
As of 10.6, the OS appears to be messing with window position and visibility in a way that was confusing the contact list hiding code. It simultaneously thought the window was both hidden and in the right position. Counteract
this by having it move to the correct position.

The if here was likely to try and guard against this being called too often, but it doesn't appear it's called outside of when necessary. Mess with it a little to control animation, not movement.

--HG--
extra : transplant_source : %DF95s%B2B%AE%ED/q%9B%AB.%7E%C8%3A%C9b%CC%E5